+++ refarch-gateway/src/test/java/de.muenchen.oss.refarch.gateway/TestConstants.java
@@ -2,7 +2,7 @@
  * Copyright (c): it@M - Dienstleister für Informations- und Telekommunikationstechnik
  * der Landeshauptstadt München, 2024
  */
-package @muenchen;
+package de.muenchen.oss.refarch.gateway;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
